refix review comments
