+++ backend/api/appsettings.json
@@ -9,7 +9,7 @@
     "CallbackPath": "/signin-oidc"
   },
   "ApplicationInsights": {
-    "InstrumentationKey": "6f08fb54-d348-44cb-9fa6-be2c9e29d419"
+    "InstrumentationKey": ""
   },
   "Logging": {
     "LogLevel": {
