+++ frontend/src/App.tsx
@@ -4,22 +4,6 @@ import { useCurrentContext, useCurrentUser } from '@equinor/fusion'
 import { Switch, Route } from 'react-router-dom'
 import ProjectRoute from './routes/ProjectRoute'
 import EvaluationRoute from './routes/EvaluationRoute'
-import { ApplicationInsights } from '@microsoft/applicationinsights-web'
-import { ReactPlugin, withAITracking } from '@microsoft/applicationinsights-react-js'
-import { createBrowserHistory } from "history"
-
-const browserHistory = createBrowserHistory({ basename: '' })
-const reactPlugin = new ReactPlugin()
-const appInsights = new ApplicationInsights({
-    config: {
-        instrumentationKey: '6f08fb54-d348-44cb-9fa6-be2c9e29d419',
-        extensions: [reactPlugin],
-        extensionConfig: {
-            [reactPlugin.identifier]: {history: browserHistory}
-        }
-    }
-})
-appInsights.loadAppInsights()
 
 const App = () => {
     const currentProject = useCurrentContext()
@@ -44,4 +28,4 @@ const App = () => {
     </>
 }
 
-export default withAITracking(reactPlugin, App)
+export default App
