+++ frontend/src/config.ts
@@ -4,6 +4,7 @@ declare const COMMITHASH: string
 declare const BRANCH: string
 declare const LAST_COMMIT: string
 declare const AD_APP_ID: string
+declare const APP_INSIGHTS: string
 
 export const config = {
     API_URL: API_URL,
@@ -11,5 +12,6 @@ export const config = {
     COMMITHASH: COMMITHASH,
     BRANCH: BRANCH,
     LAST_COMMIT: LAST_COMMIT,
-    AD_APP_ID: AD_APP_ID
+    AD_APP_ID: AD_APP_ID,
+    APP_INSIGHTS: APP_INSIGHTS
 }
